Add base GitHub Actions configuration
Adapted from the version of alphagov/pay-webhooks, and including
configuration to only build this portion of the monorepo.

As we're specifying the `paths` that are to be built, we need to make
sure that we've also got the Workflow itself, so we can make sure that
any changes to them are triggered before any code changes are required.

We make sure to include the Workflows that are in scope for this