Enable SonarQube scanning
To ensure that we're consistently delivering great code, we should
utilise SonarQube quality scanning, using SonarCloud.

We need to make sure we don't perform a shallow clone, as it allows us
to perform better scanning, and make sure that we cache our Sonar
packages.

As this is a monorepo, we've set up the monorepo style naming in the
projectKey.