Create base implementation for endpoint for retrieving all apis
As part of our proof of concept we want to create a metadata model that we
expose as the producer, but will also want to consume it ourselves later on.

To start off with we want to produce the metadata model by creating an
endpoint. We haven’t added Spring annotations yet so that we can concentrate on
unit tests to check the behaviour of the controller and the service without the
Spring context.

This also requires that we implement a basic service layer.

Co-authored-by: Jamie Tanna <jamie@jamietanna.co.uk>