Add an /apis endpoint
At this point we’ve added Spring annotations to our classes so that we can
start the Spring application context to test the endpoint.

This also requires that we implement the MetadataRepository interface, with a
no-op implementation.

We've made this return an empty, unmodifiable list to make it clear that there
is nothing in the repository.