Add a JSON schema for the API metadata model
To improve the usability of the API Metadata model outside of just our
example Java project, we should produce a language independent markup
for the structure of objects.

JSON Schema is the best solution for this, and we can use `draft-07` as
a well implemented version across different languages.

This introduces a schema for the current state of the objects, as seen
in `examples/java`:

- the API metadata object itself
- the response object for many API metadata results

For both types of objects, we want to allow consumers and producers of
the metadata model to produce extra metadata to inform future decisions
around metadata that could be made part of a future version of the
specification, so want to mark `additionalproperties` explicitly as
`true`, instead of leaving it implicitly `true`.

This also includes `ajv` to validate that the schema is well-formed JSON
Schema, and requires `ajv-formats` to validate that the `format` and
`pattern` fields are correctly formed.

We've set up separate GitHub Actions for `push` and `pull_request`s so
we can split them out in the future, to i.e. publish a package on
`main`.