Use the objects generated by the json schemas in the production code
We've used the jsonschema2pojo plugin to generate the objects. Because it's not
an official Gradle plugin we've had to use buildscript.

The DAO is always going to return a String so in the service class we wrapped
the `url` and `documentationUrl` fields in `URI.create` as in the schema we
have specified that the type is a URI. We also had to refactor the tests to do
this.

The DAO also returns the `apiVersion` as a String, whereas in the schema we've
specified the `api-version` as an enum value, so in the service class and tests
we've also used the `fromValue` method to get the enum value from the String.
We used this method because we wanted to pass the readable friendly version
rather than the enum name itself.

In the service test we removed some code that was testing different API
versions were being set, as the enum only contains a single value at the moment
so is always the same. We aso replaced the mock dao with a real dao as we would
have had to do a lot of set up on the mock to get it to work and using a real
one was easier.

In the controller test we also changed one of the tests to check the name was
set correctly instead of the version, again because the version is the same
each time and we are already testing this in the service class.

We also changed one of the BulkMetadataResponse tests to assert that null is
returned if the response is null, rather than an empty String, as this class is
now generated so don't have control over its contents. We previously had some
logic in there to intialise an empty list if null was given.

We've renamed the package where we generate the classes from json.  The schemas
are kept outside of this `spring-boot` project because we want them to be
language independent, and we want this to be clear in the package naming.