Make all Gradle builds reproducible
Reproducible builds is one of the things for the highest level of [Supply-chain
Levels for Software Artifacts
(SLSA)](https://slsa.dev/requirements#reproducible).

One advantage of this is security - because the output is always the same if
the source code doesn't change, you know that nothing has been modified to
inject vulnerabilities. This feature is easy to add in Gradle so worth doing.