Add a central OpenAPI document
We want to document the API in both a human and machine readable way, which
OpenAPI allows us to do. This is a central document that others can use to
integrate / implement themselves.

We're using Stoplight's Github action to validate the OpenAPI document, as
longer term we might want to use Spectral anyway and we can also use this to
validate the spec too.

We've had to amend our use of the `jsonSchema2Pojo` plugin in Gradle and tell
it to only include json files in the `schemas/v1alpha` directory so that it
doesn't pick up the openapi.yml file and try to turn it into a Java object.