Upgrade to draft2019 of JSON Schema
Although we'd wanted to use draft-07 as it's very well supported, we
should look to upgrade to a slightly newer version.

Draft 2019-09 is still fairly well supported, and introduces better
validation built into the standard for consumers of the validation.

This requires we update the `$schema` URIs, as well as force `ajv` to
use this specific schema version as it does not automagically detect it.