Validate that /apis matches JSON schema
As noted in 469762e8a2af3a8c0b96e993df2c874bdc252570, we found a case
where our API's response didn't actually match the JSON Schema that we
said it was, as we didn't have quite enough validation available.

To avoid hitting common issues in the future, we can ensure that we
validate that the response from the API validates against the schema.

This requires we:

- make the schemas available for the tests' runtime classpath, by
  copying them into the build directory's test resources directory, only
  after `processTestResources` has executed (using `finalizedBy`)
- utilise Rest Assured's JSON Schema Validator, as it has a handy
  Hamcrest matcher that we can use together with Spring's
  `content().string()` method
- set up the test with correctly formed `ApiMetadata` objects