Deploy the spring boot app to PaaS
The default Java version for the buildpack is Java 8 so we need to explicitly
set it to Java 17.

In the deploy step in the Github workflow we need to specify that we're using
the sandbox environment as this is where we are storing the secrets we need.

We're pinning the version of the Java buildpack to the current latest version,
so that if there are breaking changes in the future they won't break our
workflow.

We also decided to explicitly name the jar file we're building so that it
doesn't include the version number, as this means we can use the name of the
jar file in the push command without having to update the action every time we
update the version.

We don't need to checkout the repo in the `deploy` job, as we only need the jar
file and the manifest file to deploy to PaaS, which are now both uploaded in
the previous `test` job.