Use `spectral-ruleset-govuk-public` for API linting
As we now have a Spectral ruleset published to NPM, we should start
using it in this project to make sure that we're continually using it,
rather than using the out-of-the-box Spectral OpenAPI rules.

This requires that we pin the ruleset using our `package.json`.

We also need to amend our OpenAPI's `version` to match the rules
required by the ruleset.

Until the spectral-cli is upgraded (as noted in
https://github.com/stoplightio/spectral/issues/2019) we need to
explicitly pull the `spectral-ruleset-migrator` version that includes
the fix for resolving NPM dependencies correctly.