docs: Change url values to IETF RFC 3986
IETF RFC 3986 is largely more stringent than the WHATWG one which helps
reduce complexity of what is acceptable or not in the system.

Some other constraints have been added on top of that that should make
the system less prone to security exploits whilst keeping it as a valid
subset of the RFC to be compatible with most URL parsers out there.

Extra rules can be applied either before of after the initial parsing.