Add a filter for correlation-id
If a correlation-id is provided, we want it to be added to the
response and our logs. If one isn't provided we would generate one.

We've added some tests to verify that the filter is used and that the
correlation-id is added to the response even if there is an error with the filter.