Don't return HTTP objects from Service layer
While starting to migrate the application's architecture to a
Layered/Onion architecture using jMolecules, we noticed that we've got
HTTP layer objects (`ApiMetadata`) being returned by `MetadataService`.

Because we'll be moving between API versions, this is likely to change
in the future, or require other mappings between versions, as well as
being a poor architectural choice, as it leads to crossing domain
boundaries.

Instead, we use the `MetadataDto` as the object for data transfer, which
allows us to use a `record`, as it can be left immutable, and then be
remapped to other types as needed.

We can extract out the mapping between types to a `V1AlphaMapper`, which
can perform mappings between data types for the `v1alpha` specification.

We can also create a convenience method for
`MetadataDto::of(MetadataDao)` as it's something we'll likely want to do
in a few places more easily.