Migrate to NPM task for Spectral linting
Instead of using the GitHub Action, which only works when pushing to
GitHub and also requires we use an older version of the Spectral CLI, we
can instead migrate to using the CLI to lint our specification.

This requires we add the Spectral CLI to the `package.json`, so we can
correctly run it.