Don't return DAO objects from `MetadataRepository`
As we're about to migrate to Onion architecture, we don't want our
`MetadataRepository`, in the Infrastructure layer to be consumed by the
`MetadataService`,  in the Application layer, and providing
`MetadataDao` objects, as they're purely implementation detail for our
persistence tier.

Instead, we can replace it with the `MetadataDto` Domain objects, which
can then be consumed more easily.

As the `MetadataDao` is now unused, we can remove it, and re-add it at
the point that it makes sense.

Co-authored-by: Bethan Palmer <bethan.palmer@digital.cabinet-office.gov.uk>