Replace `Metadata` references with `Api`
Although `Metadata` was used to be short for `ApiMetadata`, we are
leaving it unclear by just calling it `Metadata`.

Instead of prefixing it with `ApiMetadata`, we can instead think about
this in terms of its place in the Domain, in which it'll be an `Api`.
The metadata about the API, such as what `url` it's at, is going to be
part of the `Api` object, so the name `Metadata` doesn't make sense to
keep.

Co-authored-by: Bethan Palmer <bethan.palmer@digital.cabinet-office.gov.uk>