Require Domain ring is only POJOs
The domain ring should be very lightweight, allowing us to implement
using as pure-as-possible Java, as it's unlikely to be using external
requirements given it's a key part of our application.

To capture this, we'll ad an ArchUnit test to allow use of core Java
features, and the domain ring itself for implementation.

Co-authored-by: Bethan Palmer <bethan.palmer@digital.cabinet-office.gov.uk>