Add a correlation id to our contract
We want to make sure we're adding the correlation-id to each request and response,
so that we can track a request across a whole interaction with our service.

We've explicitly set required to false (this is the default), as at the moment
this would make it harder e.g. to curl the endpoint or go to it in the browser.
We might want to make this required later on.

We've specificed that the format must be uuidv4, as version 4 is a commonly
used version that has a low chance of collision, so have added some tests to
validate that only version 4 is accepted.

The ID will also be included in the response header, so that consumers will
also have a view of it, if they had not provided one in the request.

Because the correlation-id is optional, we need to make sure we generate one if none
is provided.

We've added some tests to verify that the filter is used and that the
correlation-id is added to the response even if there is an error with the
filter.