Add error handling for the correlation-id filter
We need to be able to handle errors correctly as part of the filter.

This may be different depending on the version, so at this point we are also
introducing content negotiation to handle errors for different versions.

We're using a library for the actual content negotiation as this removes the
need for some boilerplate code.

In the decorator class we're updating the response in a separate method,
because as we add more versions as we go the decorateWithNegotiation method
would end up getting very big.

We are excluding the filter in our MetadataControllerIntegration test because
we want this test class to focus on the controller's functionality, and we have
a separate integration test for the filter itself.