Migrate ArchUnit rules to method-based rules
As IntelliJ's not (yet) got support for the field-based rules being
tests themselves, we have a not great developer experience, resulting in
us needing to run the whole test class, rather than re-run a single
rule.

We can amend this by replacing it with a method-based rule setup, which
requires we add the `.check()` call on an injected `JavaClasses` object.

Closes #76.