Move plugin declarations to `buildSrc`
We're seeing that Dependabot is bumping the version of Spotless in two
separate PRs - one for `buildSrc` and one for the main project.

This is a bit awkward, and can be hopefully minimised by putting the
plugins into `buildSrc`'s `implementation` classpath.

This then means that versions of plugins can be controlled in a single
place.