Add a scheduled method
We're using the Spring `@Scheduled` annotation to run a task once an hour.

We've parameterised the rate for the schedule so that we can set it to a
different value for the integration test.  We're setting it to  half a second
instead of an hour for this, so that it only has to wait one second for it to
have executed. There is a library called Awaitility we can use to test this.

We've also added some logging so that we can see when the scheduled task starts
and ends.