Implement the `fetchAndSaveApis` method
To implement this we've created the interfaces for `FetcherService` and
`Registry` and added empty implementations so that we can autowire them into
the `ScheduleHandler` class, which allows us to test this method.

We're saving each API after getting the list from each registry entry, so that
we don't have to wait for all lists of APIs to be fetched before we start
saving them.