Implement the `Registry` in the domain layer
As a Domain Layer object, we should be implementing it in the Domain
layer.

The means for storage of the registry entries should be injected in, and
implemented elsewhere, as it'll be very application/infrastructure
specific.

We also need to add a `DomainLayerConfiguration` class to configure any
beans that need to be created from the Domain Layer.

This also requires we move the `NoopRegistryStorage` to the
Infrastructure ring, as it's now a repository, which makes sense to be
implemented in the Infrastructure, closer to the longer term goal of an
external datastore.