Implement the `FetcherService`
As part of the first step towards being able to implement the fetching
layer across services, we need to implement the `FetcherService` in the
Domain ring.

This requires we define the core algorithm for how we want to handle
errors, requiring we split into three individual exceptions for the core
problems that may occur.

We don't bother doing anything with the exceptions thrown in
`ScheduleHandler`, expecting that they'll be logged elsewhere.

Instead of abstracting the logging functionality into a domain-layer
`Logger` class, we allow SLF4J, as it simplifies the work we need to
perform, and it's unlikely to change over time.