Implement a `Fetcher` for the V1alpha contract
To be able to fully pull from other implementations of the contract, we
need to implement the `Fetcher` in the Infrastructure ring, as it's tied
to HTTP.

To simplify things, we can use a Unit Integration test to validate that
the client works, instead of a mix of Unit and Unit Integration.

We also add some operational logging to improve supportability.

We make sure to pass the `correlation-id` from the MDC, so requests are
traced end-to-end, as well as making sure the MDC contains both request
and response `correlation-id` for diagnosing.

We split our nested integration test classes into their own spring
contexts, so we don't have a clash between `MockWebServer`s.

We can do this using the `WebClient`, which is a modern alternative to
the `RestTemplate` traditionally used, and has benefits around how we
handle errors.

We make our client as resilient to known errors as we can by handing
cases that could be misimplementation of the contract.