Generate a `correlation-id` if not present in the MDC
In cases we are being called by the `ScheduledHandler`, we will not have
a `correlation-id` in the MDC, as there's no ongoing request that will
have one generated.

Instead of requiring consumers to always set one, we can instead make it
part of this class' contract to always generate one if not present.