Run the openapi-docs workflows when npm packages change
When a new version of spectral cli or the gov.uk spectral ruleset is released,
we should lint our openapi document to check it passes any new rules that have
been added. Therefore we should run this workflow when the version of either of
these packages changes in `package.json`.